+++ forui/lib/src/theme/theme.dart
@@ -98,11 +98,6 @@ class FTheme extends ImplicitlyAnimatedWidget {
   /// The platform variant. Defaults to the current platform.
   final FPlatformVariant? platform;
 
-  /// The duration for which subsequent [FTooltip]s will appear instantly after one is shown. Defaults to 300ms.
-  ///
-  /// See [FTooltipGroup.activeDuration] for more details.
-  final Duration tooltipGroupActiveDuration;
-
   /// The widget below this widget in the tree.
   final Widget child;
 
@@ -112,7 +107,6 @@ class FTheme extends ImplicitlyAnimatedWidget {
     required this.child,
     this.textDirection,
     this.platform,
-    this.tooltipGroupActiveDuration = const Duration(milliseconds: 300),
     this.motion = const FThemeMotion(),
     super.onEnd,
     super.key,
@@ -128,8 +122,7 @@ class FTheme extends ImplicitlyAnimatedWidget {
       ..add(DiagnosticsProperty('motion', motion))
       ..add(DiagnosticsProperty('data', data))
       ..add(EnumProperty('textDirection', textDirection))
-      ..add(DiagnosticsProperty('platform', platform))
-      ..add(DiagnosticsProperty('tooltipGroupActiveDuration', tooltipGroupActiveDuration));
+      ..add(DiagnosticsProperty('platform', platform));
   }
 }
 
@@ -146,7 +139,6 @@ class _State extends AnimatedWidgetBaseState<FTheme> {
     data: _tween!.evaluate(animation),
     textDirection: widget.textDirection ?? Directionality.maybeOf(context) ?? .ltr,
     platform: widget.platform,
-    tooltipGroupActiveDuration: widget.tooltipGroupActiveDuration,
     child: widget.child,
   );
 }
@@ -235,9 +227,6 @@ class FBasicTheme extends StatelessWidget {
   /// The platform variant. Defaults to the current platform.
   final FPlatformVariant? platform;
 
-  /// The duration for which subsequent [FTooltip]s will appear instantly after one is shown. Defaults to 300ms.
-  final Duration tooltipGroupActiveDuration;
-
   /// The widget below this widget in the tree.
   final Widget child;
 
@@ -247,7 +236,6 @@ class FBasicTheme extends StatelessWidget {
     required this.child,
     this.textDirection,
     this.platform,
-    this.tooltipGroupActiveDuration = const Duration(milliseconds: 300),
     super.key,
   });
 
@@ -263,7 +251,7 @@ class FBasicTheme extends StatelessWidget {
             fontFamily: data.typography.defaultFontFamily,
             color: data.colors.foreground,
           ),
-          child: FTooltipGroup(activeDuration: tooltipGroupActiveDuration, child: child),
+          child: child,
         ),
       ),
     ),
@@ -275,8 +263,7 @@ class FBasicTheme extends StatelessWidget {
     properties
       ..add(DiagnosticsProperty('data', data, showName: false))
       ..add(EnumProperty('textDirection', textDirection))
-      ..add(DiagnosticsProperty('platform', platform))
-      ..add(DiagnosticsProperty('tooltipGroupActiveDuration', tooltipGroupActiveDuration));
+      ..add(DiagnosticsProperty('platform', platform));
   }
 }